rename prepare_for_inference to strip
