+++ examples/tensorflow/classification/main.py
@@ -294,7 +294,7 @@ def run(config):
 
     if 'export' in config.mode:
         save_path, save_format = get_saving_parameters(config)
-        export_model(compression_ctrl.prepare_for_inference(), save_path, save_format)
+        export_model(compression_ctrl.strip(), save_path, save_format)
         logger.info('Saved to {}'.format(save_path))
 
     close_strategy_threadpool(strategy)
@@ -332,7 +332,7 @@ def export(config):
                         ckpt_path=config.ckpt_path)
 
     save_path, save_format = get_saving_parameters(config)
-    export_model(compression_ctrl.prepare_for_inference(), save_path, save_format)
+    export_model(compression_ctrl.strip(), save_path, save_format)
     logger.info('Saved to {}'.format(save_path))
