+++ examples/tensorflow/object_detection/main.py
@@ -389,7 +389,7 @@ def run(config):
 
     if 'export' in config.mode:
         save_path, save_format = get_saving_parameters(config)
-        export_model(compression_ctrl.prepare_for_inference(), save_path, save_format)
+        export_model(compression_ctrl.strip(), save_path, save_format)
         logger.info('Saved to {}'.format(save_path))
 
     close_strategy_threadpool(strategy)
@@ -411,7 +411,7 @@ def export(config):
         load_checkpoint(checkpoint, config.ckpt_path)
 
     save_path, save_format = get_saving_parameters(config)
-    export_model(compression_ctrl.prepare_for_inference(), save_path, save_format)
+    export_model(compression_ctrl.strip(), save_path, save_format)
     logger.info('Saved to {}'.format(save_path))
