+++ examples/torch/classification/main.py
@@ -224,7 +224,7 @@ def main_worker(current_gpu, config: SampleConfig):
         load_state(model, model_state_dict, is_resume=True)
 
     if is_export_only:
-        export_model(compression_ctrl.prepare_for_inference(), config.to_onnx)
+        export_model(compression_ctrl.strip(), config.to_onnx)
         logger.info(f'Saved to {config.to_onnx}')
         return
 
@@ -297,7 +297,7 @@ def main_worker(current_gpu, config: SampleConfig):
     config.mlflow.end_run()
 
     if 'export' in config.mode:
-        export_model(compression_ctrl.prepare_for_inference(), config.to_onnx)
+        export_model(compression_ctrl.strip(), config.to_onnx)
         logger.info(f'Saved to {config.to_onnx}')
