+++ examples/torch/classification/staged_quantization_worker.py
@@ -220,7 +220,7 @@ def staged_quantization_main_worker(current_gpu, config):
     log_common_mlflow_params(config)
 
     if is_export_only:
-        export_model(compression_ctrl.prepare_for_inference(), config.to_onnx)
+        export_model(compression_ctrl.strip(), config.to_onnx)
         logger.info(f'Saved to {config.to_onnx}')
         return
 
@@ -241,7 +241,7 @@ def staged_quantization_main_worker(current_gpu, config):
         validate(val_loader, model, criterion, config)
 
     if 'export' in config.mode:
-        export_model(compression_ctrl.prepare_for_inference(), config.to_onnx)
+        export_model(compression_ctrl.strip(), config.to_onnx)
         logger.info(f'Saved to {config.to_onnx}')
