+++ examples/torch/object_detection/main.py
@@ -210,7 +210,7 @@ def main_worker(current_gpu, config):
     log_common_mlflow_params(config)
 
     if is_export_only:
-        export_model(compression_ctrl.prepare_for_inference(), config.to_onnx)
+        export_model(compression_ctrl.strip(), config.to_onnx)
         logger.info(f'Saved to {config.to_onnx}')
         return
 
@@ -269,7 +269,7 @@ def main_worker(current_gpu, config):
                     write_metrics(mAp, config.metrics_dump)
 
     if 'export' in config.mode:
-        export_model(compression_ctrl.prepare_for_inference(), config.to_onnx)
+        export_model(compression_ctrl.strip(), config.to_onnx)
         logger.info(f'Saved to {config.to_onnx}')
