+++ examples/torch/semantic_segmentation/main.py
@@ -539,7 +539,7 @@ def main_worker(current_gpu, config):
     log_common_mlflow_params(config)
 
     if is_export_only:
-        export_model(compression_ctrl.prepare_for_inference(), config.to_onnx)
+        export_model(compression_ctrl.strip(), config.to_onnx)
         logger.info(f'Saved to {config.to_onnx}')
         return
 
@@ -596,7 +596,7 @@ def main_worker(current_gpu, config):
         test(val_model, val_loader, criterion, color_encoding, config)
 
     if 'export' in config.mode:
-        export_model(compression_ctrl.prepare_for_inference(), config.to_onnx)
+        export_model(compression_ctrl.strip(), config.to_onnx)
         logger.info(f'Saved to {config.to_onnx}')
