+++ nncf/tensorflow/algorithm_selector.py
@@ -62,7 +62,7 @@ class NoCompressionAlgorithmController(BaseCompressionAlgorithmController):
     def statistics(self, quickly_collected_only: bool = False) -> NNCFStatistics:
         return NNCFStatistics()
 
-    def prepare_for_inference(self, do_copy: bool = True) -> tf.keras.Model:
+    def strip(self, do_copy: bool = True) -> tf.keras.Model:
         model = self.model
         if do_copy:
             model = copy_model(self.model)
