+++ nncf/torch/quantization/strip.py
