+++ tests/tensorflow/quantization/test_strip.py
@@ -20,7 +20,7 @@ from tests.tensorflow.helpers import get_basic_two_conv_test_model
 from tests.tensorflow.quantization.utils import get_basic_quantization_config
 
 
-def test_prepare_for_inference():
+def test_strip():
     model = get_basic_two_conv_test_model()
     config = get_basic_quantization_config()
     config["compression"] = {
@@ -38,7 +38,7 @@ def test_prepare_for_inference():
     input_tensor = tf.ones([1, 4, 4, 1])
     x_nncf = compressed_model(input_tensor)
 
-    inference_model = compression_ctrl.prepare_for_inference()
+    inference_model = compression_ctrl.strip()
     x_tf = inference_model(input_tensor)
 
     TFTensorListComparator.check_equal(x_nncf, x_tf)
@@ -53,7 +53,7 @@ def test_do_copy(do_copy):
         "preset": "mixed",
     }
     compression_model, compression_ctrl = create_compressed_model_and_algo_for_test(model, config, force_no_init=True)
-    inference_model = compression_ctrl.prepare_for_inference(do_copy=do_copy)
+    inference_model = compression_ctrl.strip(do_copy=do_copy)
 
     if do_copy:
         assert id(inference_model) != id(compression_model)
