+++ tests/tensorflow/sparsity/test_strip.py
@@ -21,7 +21,7 @@ from tests.tensorflow.helpers import get_empty_config
 
 
 @pytest.mark.parametrize("enable_quantization", (True, False), ids=("with_quantization", "no_quantization"))
-def test_prepare_for_inference(enable_quantization):
+def test_strip(enable_quantization):
     input_shape = (1, 4, 4, 1)
     model = get_basic_conv_test_model()
     config = get_empty_config(input_sample_sizes=input_shape)
@@ -45,7 +45,7 @@ def test_prepare_for_inference(enable_quantization):
     input_tensor = tf.ones(input_shape)
     x_nncf = compressed_model(input_tensor)
 
-    inference_model = compression_ctrl.prepare_for_inference()
+    inference_model = compression_ctrl.strip()
     x_tf = inference_model(input_tensor)
 
     TFTensorListComparator.check_equal(x_nncf, x_tf)
@@ -63,7 +63,7 @@ def test_do_copy(do_copy, enable_quantization):
         config["compression"].append({"algorithm": "quantization", "preset": "mixed"})
 
     compression_model, compression_ctrl = create_compressed_model_and_algo_for_test(model, config, force_no_init=True)
-    inference_model = compression_ctrl.prepare_for_inference(do_copy=do_copy)
+    inference_model = compression_ctrl.strip(do_copy=do_copy)
 
     # Transform model for sparsity creates copy of the model in both cases
     assert id(inference_model) != id(compression_model)
