+++ tests/torch/test_pytorch_patch.py
@@ -3,13 +3,12 @@ import inspect
 import torch
 
 from nncf.config import NNCFConfig
-from nncf.torch.dynamic_graph.patch_pytorch import MagicFunctionsToPatch
-from nncf.torch.dynamic_graph.patch_pytorch import _ORIG_JIT_SCRIPT
-from nncf.torch.graph.operator_metatypes import PT_OPERATOR_METATYPES
 from nncf.torch.dynamic_graph.context import TracingContext
-from nncf.torch.dynamic_graph.trace_tensor import TracedTensor
+from nncf.torch.dynamic_graph.patch_pytorch import _ORIG_JIT_SCRIPT
+from nncf.torch.dynamic_graph.patch_pytorch import MagicFunctionsToPatch
 from nncf.torch.dynamic_graph.trace_tensor import TensorMeta
-
+from nncf.torch.dynamic_graph.trace_tensor import TracedTensor
+from nncf.torch.graph.operator_metatypes import PT_OPERATOR_METATYPES
 from tests.shared.isolation_runner import run_pytest_case_function_in_separate_process
 from tests.torch.helpers import BasicConvTestModel
 from tests.torch.helpers import create_compressed_model_and_algo_for_test
@@ -110,5 +109,5 @@ def test_jit_trace_model():
     compressed_model, compression_ctrl = create_compressed_model_and_algo_for_test(model, config)
     torch.jit.trace(compressed_model, example_inputs=torch.rand(model.INPUT_SIZE))
 
-    model = compression_ctrl.prepare_for_inference()
+    model = compression_ctrl.strip()
     torch.jit.trace(model, example_inputs=torch.rand(model.INPUT_SIZE))
